Add new pipeline supervisor
The pipelines are now executed real-time instead of being constructed completely before and then remaining static, meaning we can update existing pipelines live, which is important when we start doing subpipelines that are spawned and die as part of a larger pipeline. Other changes include automatic restart of the processes if they die because of errors, and using temporary pipes so that if a process dies the whole pipe does not need to die with it.