Move ports to use $var for named args
If you wanted to start a process that took a stream of data as input but it wasn't isolated
in its own argv slot like `parallel 'cat $file'` where parallel's first arg is 'cat $file' and so
the syntax of {"in": "file"} does not work. Instead, you specify where you want the file to subbed
and then in the command you include a "ports" section that describes what ports are passed.